+++ CHANGELOG.rst
@@ -11,6 +11,14 @@ adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.html>`_.
 `Unreleased`_
 -------------
 
+Bugfix
+~~~~~~
+
+- Modified slightly metrics computation to avoid getting nan in some cases (issue #5)
+
+`0.1.1`_ - 2022-03-18
+---------------------
+
 New
 ~~~
 
@@ -21,6 +29,7 @@ Bugfix
 
 - For pytorch, the permutation solver would go in an infinite loop when
   a nan value was present. Corrected and test added.
+- Modified slightly metrics computation to avoid getting nan in some cases (issue #5)
 
 `0.1.0`_ - 2021-10-28
 ---------------------
@@ -37,5 +46,6 @@ Bugfix
 - Removes a stray ``print`` statement from ``sdr_loss``
 
 
-.. _Unreleased: https://github.com/fakufaku/fast_bss_eval/compare/v0.1.0...master
+.. _Unreleased: https://github.com/fakufaku/fast_bss_eval/compare/v0.1.1...master
+.. _0.1.1: https://github.com/fakufaku/fast_bss_eval/compare/v0.1.0...v0.1.1
 .. _0.1.0: https://github.com/fakufaku/fast_bss_eval/compare/v0.0.2...v0.1.0